--- python/nodeTest.py
@@ -1,4 +1,0 @@
-from web3.auto import w3
-
-connected = w3.isConnected()
-print("connected:",connected)
