+++ test/address.json
@@ -1,0 +1,7 @@
+{
+    "WalletHub": "0x04173bCBBb33897d0d869C8511fd9896596ae777",
+    "WalletTemplateInfos": "0xdc620B4c509Ebf9907896688F8071A3d40fd3Db1",
+    "WalletInfos": "0x974302262A7e00D664ce83Ff118d579aE4F4Afc0",
+    "WalletTemplateOne": "0x54F64FfBa7a188adb18ae2b364DD62d5Ab718489",
+    "WalletAdmin": "0x3330c596d2A6b8763630702e6036d9C59Ff68752"
+}
