+++ test/python/contract.py
@@ -19,8 +19,9 @@ def init():
     wallet_hub_abi_path = path + '/build/contracts/WalletHub.json'
     template_infos_abi_path = path + '/build/contracts/WalletTemplateInfos.json'
     wallet_infos_abi_path = path + '/build/contracts/WalletInfos.json'
-
     wallet_admin_abi_path = path + '/build/contracts/WalletAdmin.json'
+    template_one_abi_path = path + '/build/contracts/WalletTemplateOne.json'
+
 
 
     contract_wallet_hub_abi = loads(open(wallet_hub_abi_path).read())['abi']
@@ -28,9 +29,12 @@ def init():
         open(template_infos_abi_path).read())['abi']
     contract_wallet_infos_abi = loads(
         open(wallet_infos_abi_path).read())['abi']
-
     contract_wallet_admin_abi = loads(
         open(wallet_admin_abi_path).read())['abi']
+    contract_template_one_abi = loads(
+        open(template_one_abi_path).read())['abi']
+
+
 
     contract_wallet_hub = w3.eth.contract(
         address=all_address["WalletHub"], abi=contract_wallet_hub_abi)
@@ -40,10 +44,12 @@ def init():
         address=all_address["WalletInfos"], abi=contract_wallet_infos_abi)
     contract_wallet_admin = w3.eth.contract(
         address=all_address["WalletAdmin"], abi=contract_wallet_admin_abi)
+    contract_template_one = w3.eth.contract(
+        address=all_address["WalletTemplateOne"], abi=contract_template_one_abi)
 
-    return contract_wallet_hub, contract_template_infos, contract_wallet_infos, contract_wallet_admin
+    return contract_wallet_hub, contract_template_infos, contract_wallet_infos, contract_wallet_admin,contract_template_one
 
 
-WalletHub, WalletTemplateInfos, WalletInfos, WalletAdmin = init()
+WalletHub, WalletTemplateInfos, WalletInfos, WalletAdmin,WalletTemplateOne = init()
 
 getWallet
